move from `ctypes/__init__.pyi` to `_ctypes.pyi` - `_CDataMeta` - `_CData` - `_SimpleCData` - `Array` - fix `stubtest_allowlist`
